requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
transformers
python-multipart